Add MCX/Crypto multi-asset-class support via contract-spec-aware risk sizing
Plain NSE/BSE equity & index options already worked across the broker layer (every
adapter's `exchange` field was always a free-form string, and OrderRouter already
took an arbitrary exchange to route on). The real gap was the risk engine: it sized
every order in whole "lots" of one tenant-wide RiskConfig.lot_size, which breaks
down completely once one symbol's lot is 100 barrels of crude oil and another's is
a fraction of one Bitcoin - there's no such thing as "one lot" of a spot crypto pair.

- New app/instruments/{models,registry}.py: a ContractSpec (asset_class, lot_size,
  tick_size, fractional) and a static registry of MCX commodities (GOLD, GOLDM,
  SILVER, SILVERM, CRUDEOIL, NATURALGAS, COPPER) and crypto pairs (BTCINR, ETHINR,
  USDTINR_CRYPTO) - standard, publicly known contract specs, not fabricated. Plain
  equity/index symbols are deliberately absent; get_contract_spec() returns None for
  them and every caller treats that as "size the equity default way".
- New AssetClass enum (EQUITY/INDEX_OPTION/COMMODITY/CRYPTO).
- RiskManager.validate_and_size takes an optional contract_spec (backward compatible
  - omitted, behavior is identical to before): a non-fractional spec floors to whole
  multiples of its own lot size instead of the tenant's; a fractional one floors to
  its own smallest quantity increment without ever rounding to a whole lot.
- Quantity widened from int to float everywhere it flows through: RiskDecision,
  Trade, TradeRecord/OrderRecord (DB, new migration), BrokerOrderRequest and every
  other broker quantity field, every adapter's modify_order() - the one real blocker
  to fractional crypto sizing, since int would have silently floored 0.01 BTC to 0.
- OrderRouter.execute and the backtest engine both look up a traded symbol's
  contract spec automatically, so callers never have to know or care.
- New GET /api/instruments (list) and GET /api/instruments/{symbol} (lookup) expose
  the registry; new InstrumentsPage.tsx renders it.
- New coindcx broker stub alongside the existing Angel One/Fyers/Dhan stubs -
  structurally wired in, but no crypto exchange adapter has real tested I/O yet
  (none of the fully-implemented brokers are crypto exchanges, and testing one needs
  exchange credentials this environment can't obtain) - the same honesty constraint
  that kept the NSE provider "real but unverified" rather than faked.
- New migration b7d4f9a3c821 widens trades.quantity/orders.quantity to Float -
  verified against real Postgres with seeded pre-existing rows: applies, preserves
  data exactly, downgrades, re-applies cleanly, zero alembic check drift.
- Verified live end-to-end with Playwright: Instruments page renders correctly, and
  a full equity paper-execute round trip still works unchanged.

Backend suite: 367 passing (up from 359), including new test_instruments_api.py and
two new risk-sizing tests (MCX lot flooring vs. tenant default, crypto fractional
sizing).

+++ backend/app/brokers/base.py
@@ -57,7 +57,7 @@ class BrokerInterface(ABC):
 
     @abstractmethod
     async def modify_order(
-        self, order_id: str, quantity: Optional[int] = None, price: Optional[float] = None,
+        self, order_id: str, quantity: Optional[float] = None, price: Optional[float] = None,
         trigger_price: Optional[float] = None, order_type: Optional[str] = None,
     ) -> BrokerOrderResponse: ...
