+++ backend/app/brokers/shoonya.py
@@ -226,7 +226,7 @@ class ShoonyaBroker(BrokerInterface):
         return BrokerOrderResponse(order_id=body["norenordno"], status="OPEN", raw=body)
 
     async def modify_order(
-        self, order_id: str, quantity: Optional[int] = None, price: Optional[float] = None,
+        self, order_id: str, quantity: Optional[float] = None, price: Optional[float] = None,
         trigger_price: Optional[float] = None, order_type: Optional[str] = None,
     ) -> BrokerOrderResponse:
         payload = {"uid": self._uid, "norenordno": order_id}
