+++ backend/app/brokers/zerodha.py
@@ -248,7 +248,7 @@ class ZerodhaBroker(BrokerInterface):
         return BrokerOrderResponse(order_id=data["order_id"], status="OPEN", raw=data)
 
     async def modify_order(
-        self, order_id: str, quantity: Optional[int] = None, price: Optional[float] = None,
+        self, order_id: str, quantity: Optional[float] = None, price: Optional[float] = None,
         trigger_price: Optional[float] = None, order_type: Optional[str] = None,
     ) -> BrokerOrderResponse:
         payload = {k: v for k, v in {
